+++ include/commands/CommandRender.h
@@ -12,9 +12,16 @@
 
 DECLARE_SIMPLE_CMD(CommandMode);
 
-DECLARE_SIMPLE_CMD_NO_HELP(CommandFog);
+class CommandRenderflag : public Command {
+public:
+    virtual std::string name();
+    virtual std::string brief();
+    virtual void printHelp();
+    virtual int execute(std::istream& args);
 
-DECLARE_SIMPLE_CMD_NO_HELP(CommandLight);
+private:
+    int stringToFlag(std::string flag);
+};
 
 #endif